[query/optimizer] move optimizer common functions to common/planners
