+++ fusequery/query/src/optimizers/mod.rs
@@ -11,13 +11,11 @@ mod optimizer_projection_push_down_test;
 
 mod optimizer;
 mod optimizer_alias_push_down;
-mod optimizer_common;
 mod optimizer_limit_push_down;
 mod optimizer_projection_push_down;
 
 pub use optimizer::IOptimizer;
 pub use optimizer::Optimizer;
 pub use optimizer_alias_push_down::AliasPushDownOptimizer;
-pub use optimizer_common::OptimizerCommon;
 pub use optimizer_limit_push_down::LimitPushDownOptimizer;
 pub use optimizer_projection_push_down::ProjectionPushDownOptimizer;
