+++ fusequery/query/src/optimizers/optimizer_projection_push_down.rs
@@ -15,12 +15,12 @@ use common_planners::AggregatorPartialPlan;
 use common_planners::ExpressionPlan;
 use common_planners::FilterPlan;
 use common_planners::PlanNode;
+use common_planners::PlanRewriter;
 use common_planners::ProjectionPlan;
 use common_planners::ReadDataSourcePlan;
 use common_planners::SortPlan;
 
 use crate::optimizers::IOptimizer;
-use crate::optimizers::OptimizerCommon;
 use crate::sessions::FuseQueryContextRef;
 
 pub struct ProjectionPushDownOptimizer {}
@@ -34,7 +34,7 @@ impl ProjectionPushDownOptimizer {
 // Recursively walk an expression tree, collecting the unique set of column names
 // referenced in the expression
 fn expr_to_column_names(expr: &ExpressionPlan, accum: &mut HashSet<String>) -> Result<()> {
-    let expressions = OptimizerCommon::expression_plan_children(expr)?;
+    let expressions = PlanRewriter::expression_plan_children(expr)?;
 
     let _expressions = expressions
         .iter()
